+++ src/onetick/py/cli/__init__.py
@@ -1,0 +1,1 @@
+from .main import otp_cli