+++ src/onetick/py/core/source.py
@@ -1707,6 +1707,7 @@ class Source:
         skip_bad_tick,
         character_present,
         primary_exch,
+        show_hidden_ticks,
     )
     from ._source.source_methods.drops import (  # type: ignore[misc]
         drop,