+++ tests/core/symbols/test_symbols_from_df.py
@@ -1,6 +1,8 @@
 import onetick.py as otp
 from onetick.py.otq import otq
+
 import pytest
+import pandas as pd
 
 
 @pytest.mark.parametrize('run_mode', ['run', 'call', 'file'])
@@ -44,3 +46,17 @@ def test_df_as_symbols_no_symbol_name(session, run_mode):
             otp.run(source_2, symbols=df)
         elif run_mode == 'file':
             otp.run(source_2.to_otq(symbols=df))
+
+
+def test_run_with_dataframe_symbols(session):
+    # PY-1489
+    t = otp.Tick(X=1)
+    t['SYMBOL_PARAMS'] = otp.raw('GET_SYMBOL_PARAMETERS()', otp.varstring)
+    df = pd.DataFrame({'SYMBOL_NAME': ['A', 'B'],
+                       'SYMBOL_PARAM': ['PARAM_A', 'PARAM_B'],
+                       'TIME': [pd.Timestamp(2022, 1, 1), pd.Timestamp(2022, 1, 2)]})
+    result = otp.run(t, symbols=df)
+    assert list(result['A']['X']) == [1]
+    assert list(result['A']['SYMBOL_PARAMS']) == ["SYMBOL_PARAM='PARAM_A',TIME='2022-01-01 00:00:00'"]
+    assert list(result['B']['X']) == [1]
+    assert list(result['B']['SYMBOL_PARAMS']) == ["SYMBOL_PARAM='PARAM_B',TIME='2022-01-02 00:00:00'"]